refactor billing components
